Hook up initial UPNP lookup in WASM
